more cleanup and simplifications
