+++ packages/data-portal-schema/src/lib/dataSchemaHelpers.ts
@@ -419,9 +419,91 @@ export function getAllAttributes(
     return Array.from(allAttributes.values());
 }
 
-// Add this helper function if it's not already in the file
-function filterOutComponentAttribute(data: DataSchemaData[]): DataSchemaData[] {
+export function filterOutComponentAttribute(
+    data: DataSchemaData[]
+): DataSchemaData[] {
     return data.filter((item) => item.attribute !== 'Component');
+}
+
+export function preloadManifestData(
+    schemaData: DataSchemaData[],
+    schemaDataById: SchemaDataById
+) {
+    schemaData = filterOutComponentAttribute(schemaData);
+    const manifestDataArray = schemaData.map((schema) => {
+        const fullId = `bts:${schema.label}` as SchemaDataId;
+        const schemaData = schemaDataById[fullId];
+
+        const requiredDependencies = (
+            schemaData.requiredDependencies || []
+        ).map((depId: string | { '@id': string }) => {
+            const depSchemaId =
+                typeof depId === 'string' ? depId : depId['@id'];
+            return schemaDataById[depSchemaId];
+        });
+
+        return {
+            [schema.label]: {
+                schemaData,
+                requiredDependencies,
+                schemaDataById,
+                manifestNames: requiredDependencies.map((dep) => dep.label),
+            },
+        };
+    });
+
+    return Object.assign({}, ...manifestDataArray);
+}
+
+export function findRelatedAttributes(
+    schemaData?: DataSchemaData,
+    dataSchemaMap?: SchemaDataById
+): DataSchemaData[] {
+    if (
+        dataSchemaMap &&
+        schemaData &&
+        schemaData.exclusiveConditionalDependencies
+    ) {
+        return schemaData.exclusiveConditionalDependencies
+            .map((dependencyAttribute) => {
+                const dependencySchema = dataSchemaMap[dependencyAttribute];
+                if (
+                    dependencySchema &&
+                    dependencySchema.parentIds.some((parentId) =>
+                        schemaData.parentIds.includes(parentId)
+                    )
+                ) {
+                    return dependencySchema;
+                }
+                return null;
+            })
+            .filter((schema): schema is DataSchemaData => schema !== null);
+    }
+    return [];
+}
+
+export function hasRelatedAttributes(
+    schemaData?: DataSchemaData,
+    dataSchemaMap?: SchemaDataById
+): boolean {
+    if (
+        dataSchemaMap &&
+        schemaData &&
+        schemaData.exclusiveConditionalDependencies
+    ) {
+        return schemaData.exclusiveConditionalDependencies.some(
+            (dependencyAttribute) => {
+                const dependencySchema = dataSchemaMap[dependencyAttribute];
+                return (
+                    dependencySchema &&
+                    dependencySchema.parentIds.some((parentId) =>
+                        schemaData.parentIds.includes(parentId)
+                    )
+                );
+            }
+        );
+    }
+    return false;
 }
 
 async function getConditionalIfData(): Promise<{
